// frontend/src/app/core/services/api/model/athleteBatchMigrationRequest.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AthleteBatchMigrationRequestItem } from './athleteBatchMigrationRequestItem';


/**
 * Request payload for batch migration of athletes from old card number format
 */
export interface AthleteBatchMigrationRequest { 
    /**
     * Year for which to create accreditations
     */
    year: number;
    /**
     * List of athletes to migrate (max 100 per request)
     */
    athletes: Array<AthleteBatchMigrationRequestItem>;
}