// frontend/src/app/core/services/api/model/athleteBatchMigrationRequestItem.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Gender } from './gender';


/**
 * Single athlete migration item with old card number
 */
export interface AthleteBatchMigrationRequestItem { 
    /**
     * Old card number (4, 5, or 6 digits). 6-digit cards are used directly without transformation.
     */
    oldCardNumber: string;
    firstName: string;
    middleName: string;
    lastName: string;
    gender: Gender;
    dateOfBirth: string;
}
export namespace AthleteBatchMigrationRequestItem {
}