// frontend/src/app/core/services/api/model/athleteBatchMigrationResponse.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AthleteBatchMigrationResponseSkippedInner } from './athleteBatchMigrationResponseSkippedInner';
import { AthleteDto } from './athleteDto';


/**
 * Response for batch athlete migration
 */
export interface AthleteBatchMigrationResponse { 
    /**
     * Successfully migrated athletes
     */
    migrated?: Array<AthleteDto>;
    /**
     * Athletes that were skipped due to errors
     */
    skipped?: Array<AthleteBatchMigrationResponseSkippedInner>;
}