// frontend/src/app/core/services/api/model/gender.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Gender; in UI display as Мъж (MALE) or Жена (FEMALE)
 */
export const Gender = {
    MALE: 'MALE',
    FEMALE: 'FEMALE'
} as const;
export type Gender = typeof Gender[keyof typeof Gender];